ajustes de bugs

+++ lib/app/pages/home/controller/home_controller.dart
@@ -8,6 +8,7 @@ import 'package:email_temporario/app/data/models/messages_model.dart'
     as messages;
 import 'package:email_temporario/app/global/helper.dart';
 import 'package:email_temporario/app/global/widgets/loading_widget.dart';
+import 'package:email_temporario/app/global/widgets/snackbar_widget.dart';
 import 'package:email_temporario/app/pages/home/repository/home_repository.dart';
 import 'package:email_temporario/app/styles/style.dart';
 import 'package:flutter/material.dart';
@@ -16,6 +17,7 @@ import 'package:flutter/services.dart';
 import '../../../global/models/account.dart';
 import '../../../global/widgets/alert_dialog_widget.dart';
 import '../../../routes/app_routes.dart';
+import 'package:flutter_native_splash/flutter_native_splash.dart';
 
 class HomeController extends GetxController {
   HomeController({
@@ -25,25 +27,28 @@ class HomeController extends GetxController {
   final HomeRepository _homeRepository;
   final emailController = TextEditingController();
   final style = Style();
-  final _helper = Helper();
+  final helper = Helper();
   AccountData account = Get.find();
 
   final _alertDlg = AlertDlg();
 
   RxString domainName = ''.obs;
   RxBool validAccount = false.obs;
-
+  RxBool emailUnread = false.obs;
   RxInt selectedBottom = 0.obs;
   RxInt totalMailsToDelete = 0.obs;
   RxList<messages.HydraMember> listEmails = <messages.HydraMember>[].obs;
 
   bool disableTimer = true;
+  bool showErrorNewEmail = true;
 
   @override
   void onReady() {
     disableTimer = true;
     getDomains();
     super.onReady();
+
+    FlutterNativeSplash.remove();
   }
 
   void createRandomEmail() {
@@ -83,6 +88,8 @@ class HomeController extends GetxController {
 
   void createAccount() async {
     disableTimer = true;
+    showErrorNewEmail = true;
+    emailUnread.value = false;
 
     if (emailController.text.trim().isEmpty) {
       return _alertDlg.alert(
@@ -96,9 +103,6 @@ class HomeController extends GetxController {
     LoadingWidget.showLoadingDialog(Get.context!);
     listEmails.clear();
 
-    // final address = '${emailController.text}${domainName.value}';
-    // const pass = '123456';
-
     final response = await _homeRepository.createAccount(
       address: emailController.text,
       pass: '123456',
@@ -111,6 +115,13 @@ class HomeController extends GetxController {
       LoadingWidget.closeDialog(Get.context!);
 
       if (validAccount.value) {
+        _alertDlg.alert(
+          title: 'Sucesso',
+          body: 'Conta de e-mail criada!',
+          context: Get.context!,
+          confirmFunction: () => Navigator.pop(Get.context!, 'Ok'),
+        );
+
         _timerGetNewEmail();
       } else {
         account.removeDataAccount();
@@ -140,29 +151,47 @@ class HomeController extends GetxController {
     if (showLoading) {
       LoadingWidget.showLoadingDialog(Get.context!);
     }
+
     final response = await _homeRepository.getAllMessages();
 
     if (response.isNotEmpty) {
+      showErrorNewEmail = true;
+      int countNewEmails = 0;
+
       final modelAllMessages = messages.MessagesModel.fromJson(response);
 
-      if (listEmails.isEmpty) {
-        listEmails.assignAll(
-          modelAllMessages.hydraMember.where((e) => !e.isDeleted),
-        );
-      } else {
-        for (var message in modelAllMessages.hydraMember) {
-          listEmails.firstWhere(
-              (element) => element.hydraMemberId == message.hydraMemberId,
-              orElse: () {
-            if (!message.isDeleted) {
-              listEmails.insert(0, message);
-            }
-            return message;
-          });
+      if (modelAllMessages.hydraMember.isNotEmpty) {
+        if (listEmails.isEmpty) {
+          listEmails.assignAll(
+            modelAllMessages.hydraMember.where((e) => !e.isDeleted),
+          );
+        } else {
+          for (var message in modelAllMessages.hydraMember) {
+            listEmails.firstWhere(
+                (element) => element.hydraMemberId == message.hydraMemberId,
+                orElse: () {
+              if (!message.isDeleted) {
+                countNewEmails++;
+                listEmails.insert(0, message);
+              }
+              return message;
+            });
+          }
+        }
+
+        _unreadEmails();
+        update(['ListViewMail']);
+
+        if (countNewEmails > 0) {
+          SnackBarWidget.show(
+            title: 'Novo email',
+            message: countNewEmails == 1
+                ? 'Um e-mail chegou em sua caixa'
+                : '${countNewEmails} novos e-mails chegaram em sua caixa',
+            icon: Icons.email,
+          );
         }
       }
-
-      update(['ListViewMail']);
 
       if (showLoading) {
         LoadingWidget.closeDialog(Get.context!);
@@ -171,14 +200,23 @@ class HomeController extends GetxController {
       if (showLoading) {
         LoadingWidget.closeDialog(Get.context!);
       }
-      _alertDlg.alert(
-        title: 'Ooops',
-        body:
-            'Não foi possível verificar novos e-mails. Tente novamente mais tarde',
-        context: Get.context!,
-        confirmFunction: () => Navigator.pop(Get.context!, 'Ok'),
-      );
+
+      if (showErrorNewEmail) {
+        showErrorNewEmail = false;
+
+        _alertDlg.alert(
+          title: 'Ooops',
+          body:
+              'Não foi possível verificar novos e-mails. Tente novamente mais tarde',
+          context: Get.context!,
+          confirmFunction: () => Navigator.pop(Get.context!, 'Ok'),
+        );
+      }
     }
+  }
+
+  _unreadEmails() {
+    emailUnread.value = listEmails.indexWhere((e) => !e.seen) >= 0;
   }
 
   showMessageDetails(messages.HydraMember showDetailsMessage) async {
@@ -203,7 +241,31 @@ class HomeController extends GetxController {
         Routes.detailsMessage,
         arguments: [message],
       )?.then(
-        (_) => _timerGetNewEmail(),
+        (result) {
+          final response = result as Map<String, dynamic>;
+
+          if (response["deleteEmail"] || response["seenEmail"]) {
+            final index = listEmails.indexWhere(
+              (e) => e.hydraMemberId == showDetailsMessage.hydraMemberId,
+            );
+
+            if (index >= 0) {
+              if (response["deleteEmail"]) {
+                listEmails.removeAt(index);
+              } else if (response["seenEmail"]) {
+                listEmails[index].seen = true;
+              }
+
+              _unreadEmails();
+              update(['ListViewMail']);
+              _timerGetNewEmail();
+            } else {
+              _timerGetNewEmail();
+            }
+          } else {
+            _timerGetNewEmail();
+          }
+        },
       );
     } else {
       _timerGetNewEmail();
@@ -237,29 +299,62 @@ class HomeController extends GetxController {
     );
   }
 
-  Future<bool> removeAndCreateNewAccount() async {
+  void removeAndCreateNewAccount() async {
     disableTimer = true;
-    LoadingWidget.showLoadingDialog(Get.context!);
 
-    final response = await _homeRepository.deleteAccount();
+    _alertDlg.confirm(
+      title: 'Confirmar',
+      body: 'Será excluído endereço de e-mail para ser criado um novo!/n'
+          'Irá levar um tempo para que esse mesmo endereço fique disponível',
+      context: Get.context!,
+      cancelFunction: () {
+        selectedBottom.value = 0;
+        Navigator.pop(Get.context!, 'Cancel');
+        _timerGetNewEmail();
+      },
+      confirmFunction: () async {
+        Navigator.pop(Get.context!, 'Ok');
 
-    if (response == 204) {
-      await account.removeDataAccount();
-      validAccount.value = false;
-    }
+        LoadingWidget.showLoadingDialog(Get.context!);
+        final response = await _homeRepository.deleteAccount();
+        LoadingWidget.closeDialog(Get.context!);
 
-    LoadingWidget.closeDialog(Get.context!);
-    return !validAccount.value;
+        if (response == 204) {
+          await account.removeDataAccount();
+          validAccount.value = false;
+          listEmails.clear();
+          emailController.text = "";
+
+          update(['ListViewMail']);
+
+          _alertDlg.alert(
+            title: 'Sucesso',
+            body: 'Conta de e-mail excluída',
+            context: Get.context!,
+            confirmFunction: () => Navigator.pop(Get.context!, 'Ok'),
+          );
+        } else {
+          _alertDlg.alert(
+            title: 'Ooops',
+            body: 'Não foi possível excluir essa conta',
+            context: Get.context!,
+            confirmFunction: () => Navigator.pop(Get.context!, 'Ok'),
+          );
+        }
+      },
+    );
   }
 
   void copyToClipboard() async {
     await Clipboard.setData(
       ClipboardData(text: '${emailController.text}${domainName.value}'),
     );
-  }
 
-  String formatDate(DateTime date) {
-    return _helper.dateFormat(date);
+    SnackBarWidget.show(
+      title: 'Copiado',
+      message: 'Endereço de e-mail copiado',
+      icon: Icons.check_circle,
+    );
   }
 
   void callSelectMailToDelete(int index) {
@@ -321,10 +416,10 @@ class HomeController extends GetxController {
     disableTimer = false;
 
     Timer.periodic(const Duration(seconds: 5), (timer) {
-      getAllMessages(showLoading: false);
-
       if (disableTimer) {
         timer.cancel();
+      } else {
+        getAllMessages(showLoading: false);
       }
     });
   }
